WordAndBannedwordMethodsCreated

+++ Services/Implements/LanguageService.cs
@@ -1,8 +1,12 @@
-﻿using Microsoft.AspNetCore.Http.HttpResults;
+﻿using AutoMapper;
+using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Tabo.DAL;
 using Tabo.DTOs.Languages;
+using Tabo.Entities;
+using Tabo.Exceptions.Languages;
+using Tabo.Exceptions.Words;
 using Tabo.Services.Abstracts;
 
 namespace Tabo.Services.Implements
@@ -10,52 +14,55 @@ namespace Tabo.Services.Implements
     public class LanguageService:ILanguageService
     {
         readonly TaboDbContext _context;
+        readonly IMapper _mapper;
 
-        public LanguageService(TaboDbContext context)
+        public LanguageService(TaboDbContext context, IMapper mapper)
         {
             _context = context;
+            _mapper = mapper;
         }
+
 
         public async Task CreateAsync(LanguageCreateDto dto)
         {
-            await _context.Languages.AddAsync(new Entities.Language
+            if(await _context.Languages.AnyAsync(x=>x.Code==dto.Code))
             {
-                Name = dto.Name,
-                Icon = dto.Icon,
-                Code = dto.Code
-            });
+                throw new LanguageExistException();
+            }
+            await _context.Languages.AddAsync(_mapper.Map<Language>(dto));
             await _context.SaveChangesAsync();
         }
 
-        public async Task<LanguageGetDto?> GetByCodeAsync(string code)
+
+        public async Task<LanguageGetDto> GetByCodeAsync(string code)
         {
             var data=await _context.Languages.FirstOrDefaultAsync(x=>x.Code==code);
 
-            return new LanguageGetDto
-            {
-                Name=data.Name,
-                Icon=data.Icon,
-                Code=data.Code
-            };
+            return _mapper.Map<LanguageGetDto>(data);
             
         }
 
+
         public async Task<IEnumerable<LanguageGetDto>> GetAllAsync()
         {
-            return await _context.Languages.Select(x => new LanguageGetDto
-            {
-                Name = x.Name,
-                Icon = x.Icon,
-                Code = x.Code
-            }).ToListAsync();
+            var datas=await _context.Languages.ToListAsync();
+            return _mapper.Map<IEnumerable<LanguageGetDto>>(datas);
         }
+
 
         public async Task DeleteAsync(string code)
         {
             var data=await _context.Languages.FindAsync(code);
-            _context.Languages.Remove(data);
-            await _context.SaveChangesAsync();
+            if (await _context.Words.AnyAsync(x => x.LanguageCode == data.Code))
+                throw new LanguageReferencedException();
+            else
+            {
+                _context.Languages.Remove(data);
+                await _context.SaveChangesAsync();
+            }
         }
+
+
         public async Task UpdateAsync(LanguageUpdateDto dto, string code)
         {
             var data=await _context.Languages.FindAsync(code);
